refactor: move authentication to correct layer
