+++ src/data/usecases/authentication/db-authentication-protocols.ts
@@ -3,5 +3,4 @@ export * from '@/data/protocols/db/account/load-account-by-email-repository'
 export * from '@/data/protocols/criptography/hash-comparer'
 export * from '@/data/protocols/criptography/encrypter'
 export * from '@/data/protocols/update-access-token-repository'
-export * from '../authentication'
 export * from './db-authentication'
