+++ src/data/usecases/load-account-by-token/db-load-account-by-token.ts
@@ -1,7 +1,7 @@
 import { Decrypter } from '@/data/protocols/criptography/decrypter'
 import { LoadAccountByToken } from '../../../domain/usecases/load-account-by-token'
-import { AccountModel } from '../../../domain/usecases/authentication/db-authentication-protocols'
 import { LoadAccountByTokenRepository } from '@/data/protocols/db/account/load-account-by-token-repository'
+import { AccountModel } from '../add-account/db-add-account-protocols'
 
 export class DbLoadAccountByToken implements LoadAccountByToken {
   constructor(
