+++ src/main/factories/usecases/account/authentication/db-authentication-factory.ts
@@ -1,5 +1,6 @@
+import { DbAuthentication } from '@/data/usecases/authentication/db-authentication'
 import { Authentication } from '@/domain/usecases/authentication'
-import { DbAuthentication } from '@/domain/usecases/authentication/db-authentication'
+
 import { BcryptAdapter } from '@/infra/criptography/bcrypt-adapter/encrypt-adapter'
 import { JwtAdapter } from '@/infra/criptography/jwt-adapter/jwt-adapter'
 import { AccountMongoRepository } from '@/infra/db/mongodb/account/account-mongo-repository'
